+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Fix a anchos y altos de dialogs inapropiados.
 - Fix a vista previa de la selección de transparencia y ajuste poblacional.
 - fix a permisos de fuentes de usuarios estándar.
+- selección automática de redondeo
+- subida de adjuntos en api
 - mpgrid busca en todas las columnas
 - fix a error en auto-filtro de datos (backoffice)